generateWriteErrorMessage移動
